pyproject.toml
[project]
name = "scraper-filmes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bs4>=0.0.2",
    "diskcache>=5.6.3",
    "google-cloud-bigquery>=3.29.0",
    "loguru>=0.7.3",
    "pandas>=2.3.3",
    "prefect>=3.1.15",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "rich>=13.9.4",
    "sqlalchemy>=2.0.37",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "types-requests>=2.32.4.20250913",
]


[tool.pytest]
pythonpath = 'src'